requirements.txt
streamlit
crewai[google-genai]
crewai-tools
google-genai

python-dotenv
